+++ BlueprintUI/Sources/ViewDescription/LayoutTransition.swift
@@ -21,13 +21,13 @@ public enum LayoutTransition {
     case none
 
     /// Layout changes will always animate with the given attributes.
-    case specific(AnimationAttributes)
+    case specific(AnimationAttributes = .default)
 
     /// The view will only animate layout changes if an inherited transition exists.
     case inherited
 
     /// The view will animate along with an inherited transition (if present) or the specified fallback attributes.
-    case inheritedWithFallback(AnimationAttributes)
+    case inheritedWithFallback(AnimationAttributes = .default)
 
 }